requirements.txt
torch==1.5.0
numpy==1.18.5
torchvision==0.6.0
tqdm==4.46.1
Pillow==8.2.0